legacy_v1/local_serve/requirements.txt
fastapi
uvicorn
requests
pydantic
PyYAML
ollama
google-genai
